# src/api/routes/v2/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"PriceDataPoint": {
				"description": "Price data point in time series",
				"properties": {
					"timestamp": {
						"type": "string",
						"description": "ISO 8601 timestamp for this data point"
					},
					"avg_price": {
						"type": "number",
						"format": "double",
						"description": "Average price during this time period"
					},
					"min_price": {
						"type": "number",
						"format": "double",
						"description": "Minimum price during this time period"
					},
					"max_price": {
						"type": "number",
						"format": "double",
						"description": "Maximum price during this time period"
					},
					"volume": {
						"type": "number",
						"format": "double",
						"description": "Number of transactions/listings during this time period"
					},
					"quality_tier": {
						"type": "number",
						"format": "double",
						"description": "Optional quality tier for this data series"
					}
				},
				"required": [
					"timestamp",
					"avg_price",
					"min_price",
					"max_price",
					"volume"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetPriceHistoryResponse": {
				"description": "Response for price history endpoint",
				"properties": {
					"game_item_id": {
						"type": "string",
						"description": "Game item UUID"
					},
					"game_item_name": {
						"type": "string",
						"description": "Game item name"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/PriceDataPoint"
						},
						"type": "array",
						"description": "Time series data points"
					},
					"start_date": {
						"type": "string",
						"description": "Start date of the time range"
					},
					"end_date": {
						"type": "string",
						"description": "End date of the time range"
					},
					"interval": {
						"type": "string",
						"description": "Time interval used for aggregation"
					}
				},
				"required": [
					"game_item_id",
					"game_item_name",
					"data",
					"start_date",
					"end_date",
					"interval"
				],
				"type": "object",
				"additionalProperties": false
			},
			"QualityTierDistribution": {
				"description": "Quality tier distribution data",
				"properties": {
					"quality_tier": {
						"type": "number",
						"format": "double",
						"description": "Quality tier (1-5)"
					},
					"quantity_available": {
						"type": "number",
						"format": "double",
						"description": "Total quantity available for this tier"
					},
					"listing_count": {
						"type": "number",
						"format": "double",
						"description": "Number of listings offering this tier"
					},
					"avg_price": {
						"type": "number",
						"format": "double",
						"description": "Average price for this tier"
					},
					"min_price": {
						"type": "number",
						"format": "double",
						"description": "Minimum price for this tier"
					},
					"max_price": {
						"type": "number",
						"format": "double",
						"description": "Maximum price for this tier"
					},
					"seller_count": {
						"type": "number",
						"format": "double",
						"description": "Number of unique sellers offering this tier"
					}
				},
				"required": [
					"quality_tier",
					"quantity_available",
					"listing_count",
					"avg_price",
					"min_price",
					"max_price",
					"seller_count"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetQualityDistributionResponse": {
				"description": "Response for quality distribution endpoint",
				"properties": {
					"game_item_id": {
						"type": "string",
						"description": "Game item UUID"
					},
					"game_item_name": {
						"type": "string",
						"description": "Game item name"
					},
					"distribution": {
						"items": {
							"$ref": "#/components/schemas/QualityTierDistribution"
						},
						"type": "array",
						"description": "Distribution data by quality tier"
					},
					"total_quantity": {
						"type": "number",
						"format": "double",
						"description": "Total quantity across all tiers"
					},
					"start_date": {
						"type": "string",
						"description": "Start date of the time range (if filtered)"
					},
					"end_date": {
						"type": "string",
						"description": "End date of the time range (if filtered)"
					}
				},
				"required": [
					"game_item_id",
					"game_item_name",
					"distribution",
					"total_quantity"
				],
				"type": "object",
				"additionalProperties": false
			},
			"QualityTierSales": {
				"description": "Sales data by quality tier",
				"properties": {
					"quality_tier": {
						"type": "number",
						"format": "double",
						"description": "Quality tier (1-5)"
					},
					"volume": {
						"type": "number",
						"format": "double",
						"description": "Total sales volume (number of items sold)"
					},
					"avg_price": {
						"type": "number",
						"format": "double",
						"description": "Average sale price for this tier"
					},
					"avg_time_to_sale_hours": {
						"type": "number",
						"format": "double",
						"description": "Average time to sale in hours"
					}
				},
				"required": [
					"quality_tier",
					"volume",
					"avg_price",
					"avg_time_to_sale_hours"
				],
				"type": "object",
				"additionalProperties": false
			},
			"QualityTierPremium": {
				"description": "Price premium data by quality tier",
				"properties": {
					"quality_tier": {
						"type": "number",
						"format": "double",
						"description": "Quality tier (1-5)"
					},
					"premium_percentage": {
						"type": "number",
						"format": "double",
						"description": "Price premium percentage compared to tier 1 baseline"
					}
				},
				"required": [
					"quality_tier",
					"premium_percentage"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GetSellerStatsResponse": {
				"description": "Response for seller stats endpoint",
				"properties": {
					"seller_id": {
						"type": "string",
						"description": "Seller ID"
					},
					"sales_by_quality": {
						"items": {
							"$ref": "#/components/schemas/QualityTierSales"
						},
						"type": "array",
						"description": "Sales data grouped by quality tier"
					},
					"inventory_distribution": {
						"items": {
							"$ref": "#/components/schemas/QualityTierDistribution"
						},
						"type": "array",
						"description": "Current inventory distribution by quality tier"
					},
					"price_premiums": {
						"items": {
							"$ref": "#/components/schemas/QualityTierPremium"
						},
						"type": "array",
						"description": "Price premium percentages by quality tier"
					}
				},
				"required": [
					"seller_id",
					"sales_by_quality",
					"inventory_distribution",
					"price_premiums"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "SC Market API v2",
		"version": "2.0.0",
		"description": "Star Citizen Market API v2 - TSOA-based with auto-generated routes and OpenAPI specification",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
...
		"/api/v2/analytics/price-history": {
			"get": {
				"operationId": "GetPriceHistory",
				"responses": {
					"200": {
						"description": "Price history time-series data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetPriceHistoryResponse"
								}
							}
						}
					}
				},
				"description": "Get price history for a game item with optional quality tier filtering\n\nTracks prices per variant over time and aggregates by quality tier.\nReturns time-series data with timestamps, average/min/max prices, and volume.\nSupports filtering by game_item_id and quality_tier.\n\nRequirements:\n- 46.1: GET /api/v2/analytics/price-history endpoint\n- 46.2: Track prices per variant over time\n- 46.3: Aggregate by quality tier\n- 46.4: Support filtering by game_item_id and quality_tier\n- 46.5: Return time-series data with timestamps, avg_price, min_price, max_price, volume",
				"summary": "Get price history",
				"tags": [
					"Analytics V2"
				],
				"security": [],
				"parameters": [
					{
						"description": "Game item UUID (required)",
						"in": "query",
						"name": "game_item_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Optional quality tier filter (1-5)",
						"in": "query",
						"name": "quality_tier",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Optional start date (ISO 8601 format, defaults to 30 days ago)",
						"in": "query",
						"name": "start_date",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Optional end date (ISO 8601 format, defaults to now)",
						"in": "query",
						"name": "end_date",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Time interval for aggregation (default: 'day')",
						"in": "query",
						"name": "interval",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"hour",
								"day",
								"week",
								"month"
							]
						}
					}
				]
			}
		},
		"/api/v2/analytics/quality-distribution": {
			"get": {
				"operationId": "GetQualityDistribution",
				"responses": {
					"200": {
						"description": "Quality tier distribution histogram data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetQualityDistributionResponse"
								}
							}
						}
					}
				},
				"description": "Get quality tier distribution for a game item\n\nShows distribution of available quality tiers with tier counts, quantities,\nprices, and seller counts. Supports filtering by game_item_id and optional\ndate range for tracking quality tier trends over time.\n\nRequirements:\n- 47.1: GET /api/v2/analytics/quality-distribution endpoint\n- 47.2: Show distribution of available quality tiers\n- 47.3: Track quality tier trends over time\n- 47.4: Support filtering by game_item_id\n- 47.5: Return histogram data with tier counts, quantities, prices, seller counts",
				"summary": "Get quality tier distribution",
				"tags": [
					"Analytics V2"
				],
				"security": [],
				"parameters": [
					{
						"description": "Game item UUID (required)",
						"in": "query",
						"name": "game_item_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Optional start date (ISO 8601 format)",
						"in": "query",
						"name": "start_date",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Optional end date (ISO 8601 format)",
						"in": "query",
						"name": "end_date",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v2/analytics/seller-stats": {
			"get": {
				"operationId": "GetSellerStats",
				"responses": {
					"200": {
						"description": "Seller analytics with sales and inventory breakdown",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetSellerStatsResponse"
								}
							}
						}
					}
				},
				"description": "Get seller analytics with sales and inventory breakdown by quality tier\n\nProvides comprehensive seller analytics including:\n- Sales volume by quality tier (from completed orders)\n- Average sale price by quality tier\n- Time-to-sale metrics by quality tier\n- Current inventory distribution by quality tier\n- Price premium percentages for higher quality tiers\n\nRequirements:\n- 48.1: GET /api/v2/analytics/seller-stats endpoint\n- 48.2: Display sales volume by quality tier\n- 48.3: Display average sale price by quality tier\n- 48.4: Display time-to-sale by quality tier\n- 48.5: Display quality tier distribution of current inventory\n- 48.6: Calculate price premium percentage for higher quality tiers",
				"summary": "Get seller analytics",
				"tags": [
					"Analytics V2"
				],
				"security": [],
				"parameters": [
					{
						"description": "Optional seller ID (defaults to current user)",
						"in": "query",
						"name": "seller_id",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	}
}